Add first draft of notes on Kubernetes
Add the first draft of notes from Kubernetes, they are in a wiki-like
structure and have been converted in Markdown for easier reading on
github.
They are not intended to be an exaustive summary of everything that I
studied or what I plan in writing in the report but just some notes that
I wrote while reading.
The links in the index file have not been fixed from my original
format yet but everything else in the others pages should work.