Add prototype of cli interface using kube-hunter
Add the skeleton of the cli that will execute all the tools.
Add the implementation for kube-hunter: it's a simple implementation
that invokes kubectl directly and extract it's output in json format in
order to manipulate it and show it at later stage.
Json has been chosen because it's a standard format that can be easily
parsed and it's the same of the output of mkit.