Add kube-bench for cisbenchmark
Add kube-bench as implementation of cis benchmark.
Add first stub of interaction with Kubernetes API directly using the
official library. I will probably transform every invokation of kubectl
in calls to the API.
The integration of kube-bench is only partial: we launch the Pod and we
execute the job but we are unable to extract the Pod name in a nice way
for retrieving the logs.