Add deployment fetching in kubesec executing
Add initial element of configuration fetching (only deployments for the
moment) to the body of the function runKubesec.
Next step is doing all other kinds of objects and then move this block
of code in a separate file and make it accept parameters.
When all the files are collected I just have to run kubesec against all
of them and collect the output.