Add basic implementation of kubesec
Add the implementation of kubesec that fetch and analyzes deployments,
replicasets, daemonsets and statefulsets. It can be extended for other
types easily due to the resources list that collect all the functions
and names.
I still have to check if some namespaced resources have to be checked
(no all_namespaces version of the API) but it's just a simple matter of
adding a loop through all namespaces that we want.