Fix fetcher for kubesec and remove kubectl usage
Comment out unsupport objects in the fetcher for kubesec.
Remove all invokations of kubectl and hardcoded waiting times, now they
are regulated by a watch loop. The only hardcoded one remains the
starting time of the pod as I have not foud yet a way to check the
status of the job while the pod is starting without triggering an error.