Redesign mkit integration
Redesign mkit integration, no more dumb invokation of run-k8s with all
its feature (node server, banner, ...). I modified the Dockerfile
(eliminate all ui dependecies), the Makefile (map the directory of the
results with the parent one and other small changes) and the scripts to
not run node and output the results in the results directory.
At the moment you get in a bash promt at the end for testing purposes
but everything should be fine. I still have to add the invokation of
make build before running in order to be sure that the image is present
in the registry.